chroot: change parameter default to false and add deprecation warning
Since the module has been defaulting to not using chroots by default,
and since the use of '-' in templates, which was triggered by using a
value of undef for the $chroot parameter, now amounts to the same thing
as setting $chroot to false, we should change the default value for this
parameter.

To give some time to users to adapt their code, we add a deprecation
warning if the $chroot parameter was explicitely set to undef.